:recycle: Separate accountTransactions  and remove unit test of HOC
