+++ src/components/accountTransactions/index.js
@@ -1,32 +1,7 @@
-import React from 'react';
+/* istanbul ignore file */
 import { connect } from 'react-redux';
 import { translate } from 'react-i18next';
-import grid from 'flexboxgrid/dist/flexboxgrid.css';
-import ExplorerTransactions from './../transactions/explorerTransactions';
-import SendTo from '../sendTo';
-import styles from './accountTransactions.css';
-
-class AccountTransactions extends React.Component {
-  // eslint-disable-next-line class-methods-use-this
-  render() {
-    return <div className={`${grid.row} ${styles.wrapper}`}>
-      <div className={`${grid['col-md-4']} ${styles.sendTo}`}>
-        <SendTo
-          address={this.props.match.params.address}
-          t={this.props.t}
-          account={this.props.account}
-          delegate={this.props.delegate}
-        />
-      </div>
-      <div className={`${grid['col-sm-12']} ${styles.transactions} ${grid['col-md-8']} ${styles.transactions}`}>
-        <ExplorerTransactions
-          history={this.props.history}
-          address={this.props.match.params.address}
-          delegate={this.props.delegate} />
-      </div>
-    </div>;
-  }
-}
+import AccountTransactions from './accountTransactions';
 
 const mapStateToProps = state => ({
   publicKey: state.account.publicKey,
